chore: move contributionsFilter method
